Create a second constant called TYPE_INVALID_REQUEST_BODY_FORMAT and set that to the string from our test: invalid_body_format.
Create a special exception class, pass it the ApiProblem object, and then have some central layer convert that into our beautiful API problem JSON formatted response. So whenever something goes wrong, we'll just need to create the ApiProblem object and then throw this special exception. That'll be it, in any situation.

When you throw an exception in Symfony - even an HttpException - it returns an HTML page. Notice the Content-Type header here of text/html. And in reality, this is returning a full, giant HTML exception page - my test helpers are just summarizing things.

For every single request, Symfony has what's called a 'request format', and it defaults to html. But there are a number of different ways to say 'Hey Symfony, the user wants json, so if something goes wrong, give them that'.