If we are in debug mode and the status code is 500, I would love for Symfony's normal exception handling to take over so we can see that big beautiful stacktrace.
In ApiExceptionSubscriber we'll need to figure out if we're in debug mode. The way to do that is to pass a  flag through the __construct method and create a property for it.

If we are in debug mode and the status code is 500 we don't want our exception subscriber to do anything.